pyproject.toml
[project]
name = "astra-agents"
version = "0.1.0"
authors = [{ name = "You" }]
description = "A fast, flexible multi-agent automation framework"
requires-python = ">=3.9"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2",
  # optional providers
  "openai>=1.40.0; extra == 'openai'",
  "google-generativeai>=0.7.2; extra == 'gemini'",
]

[project.optional-dependencies]
openai = ["openai>=1.40.0"]
gemini = ["google-generativeai>=0.7.2"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
